python-backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain dependencies
langchain==0.1.17
langchain-core==0.1.52
langchain-openai==0.1.8
langchain-community==0.0.38

# CrewAI dependencies (latest compatible versions)
crewai==0.186.1
crewai-tools==0.12.1

# HTTP requests and utilities
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data processing
pydantic==2.5.2
typing-extensions==4.8.0

# Logging and monitoring
loguru==0.7.2

# Environment and configuration
python-decouple==3.8

# OpenAI for LangChain integration
openai==1.3.0

# Authentication dependencies
supabase==2.3.0
resend==0.8.0
cryptography==41.0.7
email-validator==2.1.0